+++ src/main/java/snust/sbsp/crew/controller/CrewController.java
@@ -1,0 +1,11 @@
+package snust.sbsp.crew.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/crew")
+@RequiredArgsConstructor
+public class CrewController {
+}
